レビュー指摘対応
"a@"を正しく判定できないバグの修正
テストファイル構成変更（テスト対象とテストコードを分ける）
テストコード中のモード切替をシンボルの「定義／未定義」で行うように修正
