+++ tests/unittests/test-is_mailaddress.cpp
@@ -1,134 +1,39 @@
 ﻿#include <gtest/gtest.h>
 
-#include <cstdio>
-
 #define NOMINMAX
 #include <tchar.h>
 #include <Windows.h>
 
-// リンク依存関係を取り込まないためにダミーのマクロ定義をしておく
-//#include "debug/Debug2.cpp"
-#define debug_output(str, ...)
-#define debug_exit2(file, line, exp)
+// テスト対象関数のヘッダファイル
+//#include "util/string_ex.h" //依存関係が多いのでテスト対象の関数定義のみ抜き出し
+BOOL IsMailAddress(const wchar_t* pszBuf, int nBufLen, int* pnAddressLenfth);
 
 
-// 独自シンボル USE_STRICT_INT を解決するために参照が必要
-#include "config/build_config.h"
+// 変更前実装の定義
+// ※関数定義は IsMailAddress_20160427.cpp を参照。
+BOOL IsMailAddress_20160427(const wchar_t* pszBuf, int nBufLen, int* pnAddressLenfth);
 
-// 独自定義型 ACHAR を解決するために参照が必要
-#include "basis/primitive.h"
-
-// 独自定義型 CLogicInt を解決するために参照が必要
-// これを先にインクルードしておく必要がある
-#include "basis/SakuraBasis.h"
-
-// テスト対象関数が元々あったヘッダファイル
-#include "parse/CWordParse.h"
-
-// テスト対象関数があるcppファイルを埋め込みで取り込む
-// 他ファイルで同じファイルを取り込んではいけない
-#include "util/string_ex.cpp"
 
 // テストマクロが他と被らないようにpushする
 #pragma push_macro("TEST20180909")
 
-// このテストファイルローカルのテストモード切替フラグ
-//   0 で新旧比較モード
-//   1 で新の検証モード
-#if 0
-// 新旧動作比較用マクロ(TRUE, FALSE向け)
-#define TEST20180909(pattern, buf, len, plen) \
-	EXPECT_##pattern(IsMailAddress_20160427(buf, len, plen)); \
-	ASSERT_##pattern(IsMailAddress(buf, len, plen));
+// ローカルテストモードの定義
+// ↓をコメントインすると現新比較テストを実行できます。
+//#define REGRESSION_TEST
 
-
-// テスト対象の旧関数(コピペで埋め込み)
-#pragma region ver20160427
-
-/* 現在位置がメールアドレスならば、NULL以外と、その長さを返す
-	@date 2016.04.27 記号類を許可
-*/
-BOOL IsMailAddress_20160427( const wchar_t* pszBuf, int nBufLen, int* pnAddressLenfth )
-{
-	int		j;
-	int		nDotCount;
-	int		nBgn;
-
-
-	j = 0;
-	if( (pszBuf[j] >= L'a' && pszBuf[j] <= L'z')
-	 || (pszBuf[j] >= L'A' && pszBuf[j] <= L'Z')
-	 || (pszBuf[j] >= L'0' && pszBuf[j] <= L'9')
-	 || NULL != wcschr(L"!#$%&'*+-/=?^_`{|}~", pszBuf[j])
-	){
-		j++;
-	}else{
-		return FALSE;
-	}
-	while( j < nBufLen - 2 &&
-		(
-		(pszBuf[j] >= L'a' && pszBuf[j] <= L'z')
-	 || (pszBuf[j] >= L'A' && pszBuf[j] <= L'Z')
-	 || (pszBuf[j] >= L'0' && pszBuf[j] <= L'9')
-	 || (pszBuf[j] == L'.')
-	 || NULL != wcschr(L"!#$%&'*+-/=?^_`{|}~", pszBuf[j])
-		)
-	){
-		j++;
-	}
-	if( j == 0 || j >= nBufLen - 2  ){
-		return FALSE;
-	}
-	if( L'@' != pszBuf[j] ){
-		return FALSE;
-	}
-//	nAtPos = j;
-	j++;
-	nDotCount = 0;
-//	nAlphaCount = 0;
-
-
-	for (;;) {
-		nBgn = j;
-		while( j < nBufLen &&
-			(
-			(pszBuf[j] >= L'a' && pszBuf[j] <= L'z')
-		 || (pszBuf[j] >= L'A' && pszBuf[j] <= L'Z')
-		 || (pszBuf[j] >= L'0' && pszBuf[j] <= L'9')
-		 || (pszBuf[j] == L'-')
-		 || (pszBuf[j] == L'_')
-			)
-		){
-			j++;
-		}
-		if( 0 == j - nBgn ){
-			return FALSE;
-		}
-		if( L'.' != pszBuf[j] ){
-			if( 0 == nDotCount ){
-				return FALSE;
-			}else{
-				break;
-			}
-		}else{
-			nDotCount++;
-			j++;
-		}
-	}
-	if( NULL != pnAddressLenfth ){
-		*pnAddressLenfth = j;
-	}
-	return TRUE;
-}
-
-#pragma endregion ver20160427 definition
-
-
+// モードによってテストマクロの定義内容を変える
+#ifdef REGRESSION_TEST
+	// 新旧動作比較用マクロ(TRUE, FALSE向け)
+	#define TEST20180909(pattern, buf, len, plen) \
+		EXPECT_##pattern(IsMailAddress_20160427(buf, len, plen)); \
+		ASSERT_##pattern(IsMailAddress(buf, len, plen));
 #else
-// 新実装検証用マクロ(TRUE, FALSE向け)
-#define TEST20180909(pattern, buf, len, plen) \
-	ASSERT_##pattern(IsMailAddress(buf, len, plen));
+	// 新実装検証用マクロ(TRUE, FALSE向け)
+	#define TEST20180909(pattern, buf, len, plen) \
+		ASSERT_##pattern(IsMailAddress(buf, len, plen));
 #endif
+
+//////////////////////////////////////////////////////////////////////
 
 TEST(testIsMailAddress, CheckBlank)
 {
@@ -264,6 +169,13 @@ TEST(testIsMailAddress, CheckQuotedLocalPart)
 TEST(testIsMailAddress, CheckBadQuotedLocalPart)
 {
 	wchar_t szTest[] = L"\"test@test.com";
+	TEST20180909(FALSE, szTest, _countof(szTest) - 1, NULL);
+}
+
+// レビューコメントにより追試
+TEST(testIsMailAddress, CheckAwithAtmark)
+{
+	wchar_t szTest[] = L"a@";
 	TEST20180909(FALSE, szTest, _countof(szTest) - 1, NULL);
 }
